backend/requirements.txt
openai-whisper
tqdm
requests
regex
numpy
python-dotenv
google-genai
fastapi
uvicorn
python-dotenv
google-genai
openai-whisper
pydantic
fastapi
uvicorn
openai-whisper
google-genai
python-dotenv
pandas
openpyxl
requests
httpx
python-multipart